rename back to HelmValuesImageReplaceStepVariables so github diff plays nice
